Ajustes Archivo JSON, y ajustes al Activity

+++ app/src/main/java/com/alice403/letiturismo/SitiosTuristicosItem.kt
@@ -17,7 +17,7 @@ data class SitiosTuristicosItem(
     @SerializedName("precio")
     val precio: String,
     @SerializedName("puntuacion")
-    val puntuacion: Float,
+    val puntuacion: Double,
     @SerializedName("temperatura")
     val temperatura: String,
     @SerializedName("ubicacion")
